LPD-101541 portal-search-web: Split the paginator into two usable components
The paginator was one component that chose its own pagination from whether the total was approximate. Splitting it gives callers two they can reach for directly: ClassicSearchPaginator, which links a last page and hides the pages between behind an ellipsis, and LimitedSearchPaginator, which renders a sliding window for a total whose last page is unknown.

SearchPaginatorBar keeps what the two share, which is everything except the pagination itself: the items per page picker, the result summary, and the locale formatting of the counts. Each paginator supplies the pagination as its child.

Now that the component makes the choice, the approximate total no longer selects the window, and LimitedSearchPaginator takes moreItemsAvailable back so its forward arrow disables on the last page when the count it is handed is exact.

The Search Results widget picks between the two on the LPD-98858 feature flag, which retires the liferay-ui:search-paginator markup from this widget on both paths.
