+++ modules/apps/portal-search/portal-search-web/src/main/resources/META-INF/resources/js/index.js
@@ -9,14 +9,16 @@ export {default as CustomFilter} from './CustomFilter';
 export {default as FacetUtil} from './FacetUtil';
 export {default as Sort} from './Sort';
 
-// Components
-
 export {default as CustomConfigurationRangeOptions} from './components/CustomConfigurationRangeOptions';
-export {default as SearchPaginator} from './components/SearchPaginator';
 export {default as SelectVocabularies} from './components/SelectVocabularies';
 export {default as SortConfigurationOptions} from './components/SortConfigurationOptions';
 export {default as SystemSettingsFieldList} from './components/SystemSettingsFieldList';
 export {default as SearchBarConfigurationSuggestions} from './components/search_bar_configuration_suggestions';
+
+// Components
+
+export {default as ClassicSearchPaginator} from './components/search_paginator/ClassicSearchPaginator';
+export {default as LimitedSearchPaginator} from './components/search_paginator/LimitedSearchPaginator';
 export {default as SemanticSearchConfiguration} from './components/semantic_search_configuration';
 
 // Utils + shared
