+++ modules/apps/portal-search/portal-search-web/src/main/resources/META-INF/resources/search/results/view.jsp
@@ -86,7 +86,7 @@ SearchContainer<Document> searchContainer = searchResultsPortletDisplayContext.g
 					<c:if test="<%= paginatorReactData != null %>">
 						<div>
 							<react:component
-								module="{SearchPaginator} from portal-search-web"
+								module="{ClassicSearchPaginator} from portal-search-web"
 								props="<%= paginatorReactData %>"
 							/>
 						</div>
